added movie, update ST hierarchy
