+++ Showtime.java
@@ -2,41 +2,33 @@ import java.lang.String;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-class Showtime {
+public class Showtime {
     private static int counter = 1;
     private int id;
-    private String movie;
     private int screen;
-    private LocalDate releaseDate; //dd-mm-yyyy
-    private LocalDate earlyAccess;
+    private LocalDate showDate;
     private String time; //hh:mmam/pm
     SeatingMap seatMap;
 
-    public Showtime(String t, String m, int s, String rd){
-        this.movie = m;
+    public Showtime(int s, String sd, String t){
+        setID();
         this.screen = s;
         this.time = t;
-        setReleaseDate(rd);
-        setEarlyAccess(rd);
+        setShowDate(sd);
         this.seatMap = new SeatingMap();
-        setID();
     }
 
     //getters
-    public String getMovie() { return this.movie; }
     public int getScreen() { return this.screen; }
-    public LocalDate getReleaseDate() { return this.releaseDate; }
-    public LocalDate getEarlyAccess() { return this.earlyAccess; }
     public String getTime() {return this.time; }
     public SeatingMap getSeats() { return this.seatMap; }
     public int getID() {return this.id; }
+    public LocalDate getShowDate(){return this.showDate;}
 
     //setters
-    public void setMovie(String movie){ this.movie = movie; }
     public void setScreen(int screen){ this.screen = screen; }
-    public void setReleaseDate(String rd){ this.releaseDate = LocalDate.parse(rd, DateTimeFormatter.ofPattern("dd-MM-yyyy")); }
-    public void setEarlyAccess(String ea){ this.earlyAccess = this.releaseDate.minusWeeks(1); }
     public void setTime(String time){ this.time = time; }
+    public void setShowDate(String sd){ this.showDate = LocalDate.parse(sd, DateTimeFormatter.ofPattern("dd-MM-yyyy")); }
     public void setID(){
         this.id = counter;
         incrementCounter();
@@ -46,21 +38,5 @@ class Showtime {
         counter++;
     }
 
-    public boolean isReleased(){
-        if (this.releaseDate.isAfter(LocalDate.now())){
-            return false;
-        }
-        else {
-            return true;
-        }
-    }
 
-    public boolean isEarly(){
-        if (this.earlyAccess.isAfter(LocalDate.now())){
-            return false;
-        }
-        else {
-            return true;
-        }
-    }
 }
